更新库文件

+++ frontend/controllers/BreedinfoController.php
@@ -2,6 +2,10 @@
 
 namespace frontend\controllers;
 
+use app\models\Breed;
+use app\models\Logs;
+use app\models\User;
+use frontend\models\farmsSearch;
 use Yii;
 use app\models\Breedinfo;
 use frontend\models\breedinfoSearch;
@@ -26,7 +30,14 @@ class BreedinfoController extends Controller
             ],
         ];
     }
-
+    public function beforeAction($action)
+    {
+        if(Yii::$app->user->isGuest) {
+            return $this->redirect(['site/logout']);
+        } else {
+            return true;
+        }
+    }
 //     public function beforeAction($action)
 //     {
 //     	$action = Yii::$app->controller->action->id;
@@ -43,6 +54,9 @@ class BreedinfoController extends Controller
      */
     public function actionBreedinfoindex()
     {
+        if(Yii::$app->user->isGuest) {
+            return $this->redirect(['site/login']);
+        }
         $searchModel = new breedinfoSearch();
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
@@ -51,9 +65,26 @@ class BreedinfoController extends Controller
             'dataProvider' => $dataProvider,
         ]);
     }
-    
+
+    public function actionTemp()
+    {
+        $breeds = Breed::find()->where(['year'=>User::getYear()])->all();
+        foreach ($breeds as $breed) {
+            $farmModel = Farms::findOne($breed['farms_id']);
+            var_dump($breed['farms_id']);
+            $farmModel->isbreed = 1;
+            $farmModel->save();
+            var_dump($farmModel->getErrors());
+        }
+        echo 'finished';
+    }
+
+
     public function actionBreedinfoinfo()
     {
+        if(Yii::$app->user->isGuest) {
+            return $this->redirect(['site/login']);
+        }
     	$searchModel = new breedinfoSearch();
     	$params = Yii::$app->request->queryParams;
     	$whereArray = Farms::getManagementArea()['id'];
@@ -61,15 +92,30 @@ class BreedinfoController extends Controller
     	if (empty($params['breedinfoSearch']['management_area'])) {
     		$params ['breedinfoSearch'] ['management_area'] = $whereArray;
     	}
+//        $params['breedinfoSearch']['year'] = User::getYear();
     	$dataProvider = $searchModel->search ( $params );
     	if (is_array($searchModel->management_area)) {
     		$searchModel->management_area = null;
     	}
-    
+
+        $farmsSearch = new farmsSearch();
+        $farmsparams = Yii::$app->request->queryParams;
+        $farmsparams['farmsSearch']['isbreed'] = 1;
+        if (empty($farmsparams['farmsSearch']['management_area'])) {
+            $farmsparams ['farmsSearch'] ['management_area'] = $whereArray;
+        }
+        $farmsData = $farmsSearch->searchbreed ( $farmsparams );
+        if (is_array($farmsSearch->management_area)) {
+            $farmsSearch->management_area = null;
+        }
+        Logs::writeLogs('首页畜牧业板块');
     	return $this->render('breedinfoinfo', [
     			'searchModel' => $searchModel,
     			'dataProvider' => $dataProvider,
     			'params' => $params,
+                'farmsSearch' => $farmsSearch,
+                'farmsData' => $farmsData,
+                'farmsparams' => $farmsparams,
     	]);
     }
 
@@ -80,8 +126,13 @@ class BreedinfoController extends Controller
      */
     public function actionBreedinfoview($id)
     {
+        if(Yii::$app->user->isGuest) {
+            return $this->redirect(['site/login']);
+        }
+        $model = $this->findModel($id);
+        Logs::writeLogs('查看养殖情况',$model);
         return $this->render('breedinfoview', [
-            'model' => $this->findModel($id),
+            'model' => $model,
         ]);
     }
 
@@ -92,6 +143,9 @@ class BreedinfoController extends Controller
      */
     public function actionBreedinfocreate($farms_id)
     {
+        if(Yii::$app->user->isGuest) {
+            return $this->redirect(['site/login']);
+        }
         $model = new Breedinfo();
 
         if ($model->load(Yii::$app->request->post())) {
@@ -99,7 +153,14 @@ class BreedinfoController extends Controller
         	$model->create_at = time();
         	$model->update_at = $model->create_at;
         	$model->management_area = Farms::getFarmsAreaID($farms_id);
-        	$model->save();
+            $model->year = User::getYear();
+        	$save = $model->save();
+            if($save) {
+                $farmModel = Farms::findOne($farms_id);
+                $farmModel->isbreed = 1;
+                $farmModel->save();
+            }
+            Logs::writeLogs('新增养殖信息',$model);
             return $this->redirect(['breedinfoview', 'id' => $model->id]);
         } else {
             return $this->render('breedinfocreate', [
@@ -116,11 +177,15 @@ class BreedinfoController extends Controller
      */
     public function actionBreedinfoupdate($id)
     {
+        if(Yii::$app->user->isGuest) {
+            return $this->redirect(['site/login']);
+        }
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post())) {
         	$model->update_at = time();
         	$model->save();
+            Logs::writeLogs('更新养殖信息',$model);
             return $this->redirect(['breedinfoview', 'id' => $model->id]);
         } else {
             return $this->render('breedinfoupdate', [
@@ -137,23 +202,39 @@ class BreedinfoController extends Controller
      */
     public function actionBreedinfodelete($id)
     {
+        if(Yii::$app->user->isGuest) {
+            return $this->redirect(['site/login']);
+        }
         $this->findModel($id)->delete();
 
         return $this->redirect(['breedinfoindex']);
     }
+
+    public function actionBreedinfodeleteajax($id)
+    {
+        if(Yii::$app->user->isGuest) {
+            return $this->redirect(['site/login']);
+        }
+        $model = $this->findModel($id);
+        $model->delete();
+        Logs::writeLogs('删除养殖信息',$model);
+    }
     
     public function actionBreedinfosearch($tab,$begindate,$enddate)
     {
+        if(Yii::$app->user->isGuest) {
+            return $this->redirect(['site/login']);
+        }
     	if(isset($_GET['tab']) and $_GET['tab'] !== \Yii::$app->controller->id) {
     		if($_GET['tab'] == 'yields')
-    			$class = 'plantingstructureSearch';
+    			$class = 'plantingstructurecheckSearch';
     		else
     			$class = $_GET['tab'].'Search';
     		return $this->redirect ([$_GET['tab'].'/'.$_GET['tab'].'search',
     				'tab' => $_GET['tab'],
     				'begindate' => strtotime($_GET['begindate']),
     				'enddate' => strtotime($_GET['enddate']),
-					$class =>['management_area' =>  $_GET['management_area']],
+// 					$class =>['management_area' =>  $_GET['management_area']],
     		]);
     	} 
     	$searchModel = new breedinfoSearch();
@@ -163,6 +244,7 @@ class BreedinfoController extends Controller
 			 $_GET['enddate'] = strtotime($_GET['enddate']);
 
     	$dataProvider = $searchModel->searchIndex ( $_GET );
+        Logs::writeLogs('综合查询-养殖信息');
     	return $this->render('breedinfosearch',[
 	    			'searchModel' => $searchModel,
 	    			'dataProvider' => $dataProvider,
